+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Drop the outdated SwiftMailer dependency (#913)
 
 ### Fixed
+- Allow a `TemplateHelper` Configuration without cObj (#914)
 - Harden `AbstractDataMapper::findByPageUid` (#911)
 - Fix more PHPStan warnings (#902)